feat: introduce `LabeledTextInput` widget and replace existing UI in `MnemonicWidget` and `WatchOnlyDetailsWidget`
