+++ lib/core/widgets/mnemonic_widget.dart
@@ -1,6 +1,6 @@
 import 'package:bb_mobile/core/themes/app_theme.dart';
 import 'package:bb_mobile/core/widgets/buttons/button.dart';
-import 'package:bb_mobile/core/widgets/inputs/text_input.dart';
+import 'package:bb_mobile/core/widgets/labeled_text_input.dart';
 import 'package:bb_mobile/core/widgets/text/text.dart';
 import 'package:bip39_mnemonic/bip39_mnemonic.dart' as bip39;
 import 'package:flutter/material.dart';
@@ -125,7 +125,7 @@ class _MnemonicWidgetState extends State<MnemonicWidget> {
 
           if (widget.allowPassphrase) ...[
             const Gap(16),
-            SingleLineTextWidget(
+            LabeledTextInput(
               label: 'Passphrase',
               hint: 'Optional Passphrase',
               value: passphrase,
@@ -135,7 +135,7 @@ class _MnemonicWidgetState extends State<MnemonicWidget> {
 
           if (widget.allowLabel) ...[
             const Gap(16),
-            SingleLineTextWidget(
+            LabeledTextInput(
               label: 'Label',
               hint: 'Required',
               value: label,
@@ -508,69 +508,6 @@ class _HintChip extends StatelessWidget {
         ),
         child: Center(child: BBText(word, style: context.font.bodyLarge)),
       ),
-    );
-  }
-}
-
-class SingleLineTextWidget extends StatelessWidget {
-  final String label;
-  final String value;
-  final String hint;
-  final Function(String) onChanged;
-
-  const SingleLineTextWidget({
-    super.key,
-    required this.label,
-    required this.value,
-    required this.onChanged,
-    this.hint = '',
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        BBText(
-          label,
-          style: context.font.labelMedium?.copyWith(
-            fontWeight: FontWeight.w700,
-            color: context.colour.secondary,
-            letterSpacing: 0,
-            fontSize: 14,
-          ),
-        ),
-        const Gap(8),
-        Container(
-          decoration: BoxDecoration(
-            color: context.colour.onPrimary,
-            borderRadius: BorderRadius.circular(2.76),
-            border: Border.all(color: context.colour.surface, width: 0.69),
-            boxShadow: [
-              BoxShadow(
-                color: context.colour.surface,
-                offset: const Offset(0, 2),
-              ),
-            ],
-          ),
-          child: BBInputText(
-            value: value,
-            onChanged: onChanged,
-            style: context.font.bodySmall?.copyWith(
-              fontWeight: FontWeight.w700,
-              fontSize: 14,
-              color: context.colour.secondary,
-            ),
-            hintStyle: context.font.bodySmall?.copyWith(
-              fontWeight: FontWeight.w700,
-              fontSize: 14,
-              color: context.colour.surface,
-            ),
-            hint: hint,
-            hideBorder: true,
-          ),
-        ),
-      ],
     );
   }
 }
