+++ lib/features/import_watch_only_wallet/presentation/watch_only_details_widget.dart
@@ -2,6 +2,7 @@ import 'package:bb_mobile/core/entities/signer_device_entity.dart';
 import 'package:bb_mobile/core/themes/app_theme.dart';
 import 'package:bb_mobile/core/widgets/buttons/button.dart';
 import 'package:bb_mobile/core/widgets/inputs/text_input.dart';
+import 'package:bb_mobile/core/widgets/labeled_text_input.dart';
 import 'package:bb_mobile/core/widgets/text/text.dart';
 import 'package:bb_mobile/features/import_watch_only_wallet/presentation/cubit/import_watch_only_cubit.dart';
 import 'package:bb_mobile/features/import_watch_only_wallet/watch_only_wallet_entity.dart';
@@ -45,67 +46,77 @@ class _DescriptorDetailsWidget extends StatelessWidget {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        BBText('Descriptor', style: context.font.titleMedium),
-        const Gap(8),
-        BBText(entity.descriptor.combined, style: context.font.bodyMedium),
-        const Gap(24),
-        BBText('Type', style: context.font.titleMedium),
-        const Gap(8),
-        BBText(
-          entity.descriptor.derivation.label,
-          style: context.font.bodyMedium,
+        LabeledTextInput(
+          label: 'Descriptor',
+          value: entity.descriptor.combined,
+          onChanged: null,
         ),
         const Gap(24),
-        Row(
-          children: [
-            SizedBox(
-              width: 120,
-              child: BBText('Signing Device', style: context.font.titleMedium),
-            ),
-            SizedBox(
-              width: 200,
-              child: DropdownButtonFormField<SignerDeviceEntity?>(
-                alignment: Alignment.centerLeft,
-                decoration: const InputDecoration(
-                  border: InputBorder.none,
-                  contentPadding: EdgeInsets.symmetric(horizontal: 24.0),
+        LabeledTextInput(
+          label: 'Type',
+          value: entity.descriptor.derivation.label,
+          onChanged: null,
+        ),
+        const Gap(24),
+        if (entity.signerDevice == null)
+          Row(
+            children: [
+              SizedBox(
+                width: 120,
+                child: BBText(
+                  'Signing Device',
+                  style: context.font.titleMedium,
                 ),
-                icon: Icon(
-                  Icons.keyboard_arrow_down,
-                  color: context.colour.secondary,
-                ),
-                value: entity.signerDevice,
-                items:
-                    [null, ...SignerDeviceEntity.values]
-                        .map(
-                          (value) => DropdownMenuItem<SignerDeviceEntity?>(
-                            value: value,
-                            child: BBText(
-                              value?.displayName ?? 'Not supported',
-                              style: context.font.headlineSmall,
-                            ),
-                          ),
-                        )
-                        .toList(),
-                onChanged: cubit.onSignerDeviceChanged,
               ),
-            ),
-          ],
-        ),
+              SizedBox(
+                width: 200,
+                child: DropdownButtonFormField<SignerDeviceEntity?>(
+                  alignment: Alignment.centerLeft,
+                  decoration: const InputDecoration(
+                    border: InputBorder.none,
+                    contentPadding: EdgeInsets.symmetric(horizontal: 24.0),
+                  ),
+                  icon: Icon(
+                    Icons.keyboard_arrow_down,
+                    color: context.colour.secondary,
+                  ),
+                  value: entity.signerDevice,
+                  items:
+                      [null, ...SignerDeviceEntity.values]
+                          .map(
+                            (value) => DropdownMenuItem<SignerDeviceEntity?>(
+                              value: value,
+                              child: BBText(
+                                value?.displayName ?? 'Not supported',
+                                style: context.font.headlineSmall,
+                              ),
+                            ),
+                          )
+                          .toList(),
+                  onChanged: cubit.onSignerDeviceChanged,
+                ),
+              ),
+            ],
+          ),
+        if (entity.signerDevice != null)
+          LabeledTextInput(
+            label: 'Signing Device',
+            value: entity.signerDevice!.displayName,
+            onChanged: null,
+          ),
         const Gap(24),
-        BBText('Label', style: context.font.titleMedium),
-        const Gap(8),
-        BBInputText(
-          onChanged: cubit.updateLabel,
+        LabeledTextInput(
+          label: 'Label',
+          hint: 'Required',
           value: entity.label,
-          maxLines: 1,
+          onChanged: cubit.updateLabel,
         ),
         const Gap(24),
         BBButton.big(
           onPressed: cubit.import,
           label: 'Import',
-          bgColor: context.colour.primary,
-          textColor: context.colour.onPrimary,
+          bgColor: context.colour.secondary,
+          textColor: context.colour.onSecondary,
         ),
         const Gap(24),
       ],
